# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TaskAPI
  version: 1.0.0
  description: OpenAPI definition used for API documentation and authenticated DAST.
servers:
  - url: http://localhost:3000
paths:
...
  /api/tasks:
    get:
      operationId: listTasks
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - in: query
          name: offset
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: User task list
        "401":
          description: Authentication required
    post:
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Task created
        "400":
          description: Invalid input
        "401":
          description: Authentication required
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    TaskInput:
      type: object
      additionalProperties: false
      required: [title]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
          example: ZAP security test task
        description:
          type: string
          nullable: true
          maxLength: 5000
...